Mount /var/lock/pddf-locks in pmon (#23128)
Why I did it
In PR#22684 we added locking to the pddf led API. However, the lock is not being shared between the pmon container and the host. As a result, we have not completely eliminated these errors

[ 3919.694550] PDDF_LED ERROR find_led_ops_data invalid index: 2 for type:PSU_LED;1
[ 3919.702819] ERROR get_status_led: Cannot find LED Ptr
[ 7422.178164] PDDF_LED ERROR find_led_ops_data invalid index: 1 for type:SYS_LED;0
[ 7422.186449] PDDF_LED	ERROR get_status_led: Cannot find LED Ptr
Work item tracking
Microsoft ADO (number only):
How I did it
Mount the file lock directory (/var/lock/pddf-locks) in the pmon container. Add an intermediate directory pddf-locks so that pmon doesn't need to mount all of /var/lock.

How to verify it
Confirmed the path is mounted in pmon now. No regressions in LED control.